Split auth-provider reclaim fence tests out of the 2000-line node-test budget
Move the shared OAuth test harness into test-support and the three
fenced-reclaim cases into their own file so auth-provider.node.test.ts
stays under the slop-ratchet cap after the main merge.
