+++ packages/worker/src/app/handlers/auth-provider.node.test.ts
@@ -1,4 +1,4 @@
-import { DatabaseSync } from 'node:sqlite'
+import { type DatabaseSync } from 'node:sqlite'
 import { HttpResponse, http } from 'msw'
 import { afterAll, afterEach, beforeAll, expect, test, vi } from 'vitest'
 import { createAuthCookie, setAuthSessionSecret } from '#app/auth-session.ts'
@@ -31,12 +31,19 @@ import {
 	auditEventSummaries,
 	logAuditEventSpy,
 } from '#worker/test-support/audit-log-spy.ts'
-import { applyAllMigrations } from '#worker/test-support/apply-all-migrations.ts'
 import { emptyPublicFormProtection } from '#universal/public-form-protection.ts'
 import { reservedUsernamesKvKey } from '#worker/identity/reserved-username-settings.ts'
-import { signupModeKvKey } from '#worker/signup-mode-setting.ts'
-
-const testCookieSecret = 'test-cookie-secret-0123456789abcdef0123456789'
+import {
+	createAppEnv,
+	createMemoryKv,
+	createMigratedDb,
+	createSignupModeKv,
+	getCookiePair,
+	runHandler,
+	seedUser,
+	startProviderFlow,
+	testCookieSecret,
+} from '#worker/test-support/auth-provider-harness.ts'
 
 const msw = createMswNodeServer()
 
@@ -51,262 +58,6 @@ afterEach(() => {
 afterAll(() => {
 	msw.close()
 })
-
-function applyMigrations(db: DatabaseSync) {
-	const migrationsDir = new URL('../../../migrations/', import.meta.url)
-	applyAllMigrations(db, migrationsDir)
-}
-
-function createD1FromSqlite(db: DatabaseSync) {
-	function createPreparedStatement(query: string) {
-		return {
-			query,
-			bind(...params: Array<unknown>) {
-				return {
-					query,
-					async all<T>() {
-						const statement = db.prepare(query)
-						const rows = statement.all(...params) as Array<T>
-						return {
-							results: rows,
-							meta: { changes: 0, last_row_id: 0 },
-						}
-					},
-					async first<T>() {
-						const statement = db.prepare(query)
-						return (statement.get(...params) ?? null) as T | null
-					},
-					async run() {
-						const statement = db.prepare(query)
-						const result = statement.run(...params)
-						return {
-							meta: {
-								changes: result.changes,
-								last_row_id: Number(result.lastInsertRowid),
-							},
-						}
-					},
-				}
-			},
-			async all<T>() {
-				const statement = db.prepare(query)
-				const rows = statement.all() as Array<T>
-				return {
-					results: rows,
-					meta: { changes: 0, last_row_id: 0 },
-				}
-			},
-			async first<T>() {
-				const statement = db.prepare(query)
-				return (statement.get() ?? null) as T | null
-			},
-			async run() {
-				const statement = db.prepare(query)
-				const result = statement.run()
-				return {
-					meta: {
-						changes: result.changes,
-						last_row_id: Number(result.lastInsertRowid),
-					},
-				}
-			},
-		}
-	}
-	return {
-		prepare: createPreparedStatement,
-		async exec(query: string) {
-			db.exec(query)
-		},
-	} as unknown as D1Database
-}
-
-function withDeletingAtAfterWritableCheck(
-	db: D1Database,
-	deletingAt: string,
-): D1Database {
-	const originalPrepare = db.prepare.bind(db)
-	return {
-		...db,
-		prepare(query: string) {
-			const statement = originalPrepare(query)
-			const normalized = query.replace(/\s+/g, ' ').toLowerCase()
-			if (
-				!normalized.includes('select deleting_at from users') ||
-				!normalized.includes('stable_user_id')
-			) {
-				return statement
-			}
-			return {
-				...statement,
-				bind(...params: Array<unknown>) {
-					const bound = statement.bind(...params)
-					return {
-						...bound,
-						async first<T>() {
-							const row = await bound.first<T>()
-							await originalPrepare(
-								`UPDATE users SET deleting_at = ? WHERE stable_user_id = ?`,
-							)
-								.bind(deletingAt, params[0])
-								.run()
-							return row
-						},
-					}
-				},
-			}
-		},
-	} as D1Database
-}
-
-function withDeletingAtBeforeOauthConnectionInsert(
-	db: D1Database,
-	deletingAt: string,
-): D1Database {
-	const originalPrepare = db.prepare.bind(db)
-	return {
-		...db,
-		prepare(query: string) {
-			const statement = originalPrepare(query)
-			const normalized = query.replace(/\s+/g, ' ').toLowerCase()
-			if (!normalized.includes('insert into oauth_connections')) {
-				return statement
-			}
-			return {
-				...statement,
-				bind(...params: Array<unknown>) {
-					const bound = statement.bind(...params)
-					return {
-						...bound,
-						async run() {
-							await originalPrepare(
-								`UPDATE users SET deleting_at = ? WHERE deleting_at IS NULL`,
-							)
-								.bind(deletingAt)
-								.run()
-							return bound.run()
-						},
-					}
-				},
-			}
-		},
-	} as D1Database
-}
-
-function createMigratedDb() {
-	const sqlite = new DatabaseSync(':memory:')
-	applyMigrations(sqlite)
-	return { sqlite, db: createD1FromSqlite(sqlite) }
-}
-
-async function seedUser(
-	sqlite: DatabaseSync,
-	input: {
-		id: number
-		email: string
-		username: string
-		stableUserId?: string
-		emailVerified?: boolean
-	},
-) {
-	const passwordHash = await createPasswordHash('test-password')
-	const stableUserId =
-		input.stableUserId ?? (await createStableUserIdFromEmail(input.email))
-	const verifiedAt = input.emailVerified ? 'CURRENT_TIMESTAMP' : 'NULL'
-	sqlite.exec(`
-		INSERT INTO users (id, username, email, stable_user_id, password_hash, email_verified_at)
-		VALUES (
-			${input.id},
-			${quoteSqlString(input.username)},
-			${quoteSqlString(input.email)},
-			${quoteSqlString(stableUserId)},
-			${quoteSqlString(passwordHash)},
-			${verifiedAt}
-		);
-	`)
-}
-
-function createAppEnv(
-	db: D1Database,
-	overrides: Record<string, unknown> = {},
-): Env {
-	return {
-		APP_DB: db,
-		COOKIE_SECRET: testCookieSecret,
-		SENTRY_ENVIRONMENT: 'test',
-		SIGNUP_MODE: 'open',
-		GITHUB_CLIENT_ID: 'github-client-id-test',
-		GITHUB_CLIENT_SECRET: 'github-client-secret-test',
-		GOOGLE_CLIENT_ID: 'google-client-id-test',
-		GOOGLE_CLIENT_SECRET: 'google-client-secret-test',
-		X_CLIENT_ID: 'x-client-id-test',
-		X_CLIENT_SECRET: 'x-client-secret-test',
-		DISCORD_CLIENT_ID: 'discord-client-id-test',
-		DISCORD_CLIENT_SECRET: 'discord-client-secret-test',
-		...overrides,
-	} as unknown as Env
-}
-
-function createMemoryKv(initial?: Record<string, string>) {
-	const store = new Map<string, string>(Object.entries(initial ?? {}))
-	return {
-		async get(key: string, type?: string) {
-			const raw = store.get(key)
-			if (raw === undefined) return null
-			return type === 'json' ? JSON.parse(raw) : raw
-		},
-		async put(key: string, value: string) {
-			store.set(key, value)
-		},
-		store,
-	} as unknown as KVNamespace
-}
-
-function createSignupModeKv(mode: 'invite' | 'open' | 'waitlist') {
-	return createMemoryKv({
-		[signupModeKvKey]: JSON.stringify({
-			mode,
-			updatedAt: '2026-09-02T00:00:00.000Z',
-			updatedBy: 'admin-stable-id',
-		}),
-	})
-}
-
-type Handler = {
-	handler(context: never): Promise<Response>
-}
-
-async function runHandler(
-	handler: Handler,
-	request: Request,
-	params: Record<string, string> = {},
-): Promise<Response> {
-	return handler.handler({
-		request,
-		url: new URL(request.url),
-		params,
-	} as never)
-}
-
-function getCookiePair(setCookieHeader: string) {
-	const pair = setCookieHeader.split(';')[0]
-	if (!pair) throw new Error(`Unexpected Set-Cookie header: ${setCookieHeader}`)
-	return pair
-}
-
-async function startProviderFlow(env: Env, provider: string, url: string) {
-	const startResponse = await runHandler(
-		createAuthProviderStartHandler(env),
-		new Request(url, { method: 'POST' }),
-		{ provider },
-	)
-	expect(startResponse.status).toBe(302)
-	const location = startResponse.headers.get('Location') ?? ''
-	const stateCookie = getCookiePair(
-		startResponse.headers.get('Set-Cookie') ?? '',
-	)
-	const state = new URL(location).searchParams.get('state') ?? ''
-	return { location, stateCookie, state }
-}
 
 test('providers api lists only configured providers', async () => {
 	const { db } = createMigratedDb()
@@ -722,236 +473,6 @@ test('google sign-in reclaims an unverified password account matching the provid
 			reason: expect.stringContaining(
 				'provider=google;two_factor=1;passkeys=1;oauth_connections=1',
 			),
-		}),
-	)
-})
-
-test('google sign-in does not reclaim a fenced unverified account', async () => {
-	const { sqlite, db } = createMigratedDb()
-	const env = createAppEnv(db, {
-		OAUTH_PROVIDER: {
-			listUserGrants: async () => ({ items: [] }),
-			revokeGrant: async () => undefined,
-		},
-	})
-	await seedUser(sqlite, {
-		id: 9,
-		email: 'fenced-squat@example.com',
-		username: 'fenced-squat',
-		emailVerified: false,
-	})
-	sqlite.exec(
-		`UPDATE users SET deleting_at = '2026-09-01 12:00:00' WHERE id = 9`,
-	)
-
-	msw.use(
-		http.post('https://oauth2.googleapis.com/token', () =>
-			HttpResponse.json({ access_token: 'google-access-token' }),
-		),
-		http.get('https://openidconnect.googleapis.com/v1/userinfo', () =>
-			HttpResponse.json({
-				sub: 'google-fenced-sub',
-				email: 'fenced-squat@example.com',
-				email_verified: true,
-				name: 'Real Owner',
-			}),
-		),
-	)
-
-	const start = await startProviderFlow(
-		env,
-		'google',
-		'http://example.com/auth/google',
-	)
-	const callbackResponse = await runHandler(
-		createAuthProviderCallbackHandler(env),
-		new Request(
-			`http://example.com/auth/google/callback?code=google-auth-code&state=${start.state}`,
-			{ headers: { Cookie: start.stateCookie } },
-		),
-		{ provider: 'google' },
-	)
-	expect(callbackResponse.status).toBe(302)
-	expect(callbackResponse.headers.get('Location')).toBe(
-		'/login?oauthError=email-unavailable',
-	)
-	expect(
-		callbackResponse.headers
-			.getSetCookie()
-			.some((cookie) => cookie.startsWith('kody_session=')),
-	).toBe(false)
-
-	const user = sqlite
-		.prepare(`SELECT email_verified_at, deleting_at FROM users WHERE id = 9`)
-		.get() as { email_verified_at: string | null; deleting_at: string | null }
-	expect(user.email_verified_at).toBeNull()
-	expect(user.deleting_at).toBe('2026-09-01 12:00:00')
-	expect(
-		sqlite
-			.prepare(
-				`SELECT COUNT(*) AS count FROM oauth_connections WHERE user_id = 9`,
-			)
-			.get(),
-	).toEqual({ count: 0 })
-	expect(logAuditEventSpy).toHaveBeenCalledWith(
-		expect.objectContaining({
-			category: 'auth',
-			action: 'oauth_login',
-			result: 'failure',
-			reason: 'account_deleting',
-		}),
-	)
-})
-
-test('google sign-in does not reclaim when a purge claim lands between the writable check and the stamp', async () => {
-	const { sqlite, db: rawDb } = createMigratedDb()
-	const db = withDeletingAtAfterWritableCheck(rawDb, '2026-09-02 12:00:00')
-	const env = createAppEnv(db, {
-		OAUTH_PROVIDER: {
-			listUserGrants: async () => ({ items: [] }),
-			revokeGrant: async () => undefined,
-		},
-	})
-	await seedUser(sqlite, {
-		id: 10,
-		email: 'race-squat@example.com',
-		username: 'race-squat',
-		emailVerified: false,
-	})
-
-	msw.use(
-		http.post('https://oauth2.googleapis.com/token', () =>
-			HttpResponse.json({ access_token: 'google-access-token' }),
-		),
-		http.get('https://openidconnect.googleapis.com/v1/userinfo', () =>
-			HttpResponse.json({
-				sub: 'google-race-sub',
-				email: 'race-squat@example.com',
-				email_verified: true,
-				name: 'Real Owner',
-			}),
-		),
-	)
-
-	const start = await startProviderFlow(
-		env,
-		'google',
-		'http://example.com/auth/google',
-	)
-	const callbackResponse = await runHandler(
-		createAuthProviderCallbackHandler(env),
-		new Request(
-			`http://example.com/auth/google/callback?code=google-auth-code&state=${start.state}`,
-			{ headers: { Cookie: start.stateCookie } },
-		),
-		{ provider: 'google' },
-	)
-	expect(callbackResponse.status).toBe(302)
-	expect(callbackResponse.headers.get('Location')).toBe(
-		'/login?oauthError=email-unavailable',
-	)
-	expect(
-		callbackResponse.headers
-			.getSetCookie()
-			.some((cookie) => cookie.startsWith('kody_session=')),
-	).toBe(false)
-
-	const user = sqlite
-		.prepare(`SELECT email_verified_at, deleting_at FROM users WHERE id = 10`)
-		.get() as { email_verified_at: string | null; deleting_at: string | null }
-	expect(user.email_verified_at).toBeNull()
-	expect(user.deleting_at).toBe('2026-09-02 12:00:00')
-	expect(
-		sqlite
-			.prepare(
-				`SELECT COUNT(*) AS count FROM oauth_connections WHERE user_id = 10`,
-			)
-			.get(),
-	).toEqual({ count: 0 })
-	expect(logAuditEventSpy).toHaveBeenCalledWith(
-		expect.objectContaining({
-			category: 'auth',
-			action: 'oauth_login',
-			result: 'failure',
-			reason: 'account_deleting',
-		}),
-	)
-})
-
-test('google sign-in does not reclaim when a purge claim lands before the oauth connection insert', async () => {
-	const { sqlite, db: rawDb } = createMigratedDb()
-	const db = withDeletingAtBeforeOauthConnectionInsert(
-		rawDb,
-		'2026-09-02 12:00:00',
-	)
-	const env = createAppEnv(db, {
-		OAUTH_PROVIDER: {
-			listUserGrants: async () => ({ items: [] }),
-			revokeGrant: async () => undefined,
-		},
-	})
-	await seedUser(sqlite, {
-		id: 13,
-		email: 'race-insert-squat@example.com',
-		username: 'race-insert-squat',
-		emailVerified: false,
-	})
-
-	msw.use(
-		http.post('https://oauth2.googleapis.com/token', () =>
-			HttpResponse.json({ access_token: 'google-access-token' }),
-		),
-		http.get('https://openidconnect.googleapis.com/v1/userinfo', () =>
-			HttpResponse.json({
-				sub: 'google-race-insert-sub',
-				email: 'race-insert-squat@example.com',
-				email_verified: true,
-				name: 'Real Owner',
-			}),
-		),
-	)
-
-	const start = await startProviderFlow(
-		env,
-		'google',
-		'http://example.com/auth/google',
-	)
-	const callbackResponse = await runHandler(
-		createAuthProviderCallbackHandler(env),
-		new Request(
-			`http://example.com/auth/google/callback?code=google-auth-code&state=${start.state}`,
-			{ headers: { Cookie: start.stateCookie } },
-		),
-		{ provider: 'google' },
-	)
-	expect(callbackResponse.status).toBe(302)
-	expect(callbackResponse.headers.get('Location')).toBe(
-		'/login?oauthError=email-unavailable',
-	)
-	expect(
-		callbackResponse.headers
-			.getSetCookie()
-			.some((cookie) => cookie.startsWith('kody_session=')),
-	).toBe(false)
-
-	const user = sqlite
-		.prepare(`SELECT email_verified_at, deleting_at FROM users WHERE id = 13`)
-		.get() as { email_verified_at: string | null; deleting_at: string | null }
-	expect(user.email_verified_at).toBeNull()
-	expect(user.deleting_at).toBe('2026-09-02 12:00:00')
-	expect(
-		sqlite
-			.prepare(
-				`SELECT COUNT(*) AS count FROM oauth_connections WHERE user_id = 13`,
-			)
-			.get(),
-	).toEqual({ count: 0 })
-	expect(logAuditEventSpy).toHaveBeenCalledWith(
-		expect.objectContaining({
-			category: 'auth',
-			action: 'oauth_login',
-			result: 'failure',
-			reason: 'account_deleting',
 		}),
 	)
 })
